attendance trace

+++ app/Http/Controllers/backend/AttendanceController.php
@@ -15,7 +15,8 @@ class AttendanceController extends Controller
     {
         $this->middleware('admin');
     }
-    public function list(Request $request){
+    public function list(Request $request)
+    {
 
         $data['title'] = Config::get('constants.PROJECT_NAME') . ' || Attendance List';
         $data['description'] = Config::get('constants.PROJECT_NAME') . ' || Attendance List';
@@ -56,7 +57,8 @@ class AttendanceController extends Controller
 
 
 
-    public function dayList(Request $request){
+    public function dayList(Request $request)
+    {
 
         $objEmployee = new Employee();
         $data['employee'] = $objEmployee->get_admin_employee_details();
@@ -97,14 +99,15 @@ class AttendanceController extends Controller
         );
         return view('backend.pages.attendance.attendance_day_list', $data);
     }
-    public function add (){
+    public function add()
+    {
 
         $objEmployee = new Employee();
         $data['employee'] = $objEmployee->get_admin_employee_details();
 
-        $data['title'] = Config::get( 'constants.PROJECT_NAME' ) . " || Add Attendance";
-        $data['description'] = Config::get( 'constants.PROJECT_NAME' ) . " || Add Attendance";
-        $data['keywords'] = Config::get( 'constants.PROJECT_NAME' ) . " || Add Attendance";
+        $data['title'] = Config::get('constants.PROJECT_NAME') . " || Add Attendance";
+        $data['description'] = Config::get('constants.PROJECT_NAME') . " || Add Attendance";
+        $data['keywords'] = Config::get('constants.PROJECT_NAME') . " || Add Attendance";
         $data['css'] = array(
             'toastr/toastr.min.css'
         );
@@ -133,7 +136,8 @@ class AttendanceController extends Controller
         );
         return view('backend.pages.attendance.add', $data);
     }
-    public function saveAdd(Request $request){
+    public function saveAdd(Request $request)
+    {
         $objAttendance = new Attendance();
         $result = $objAttendance->saveAdd($request);
 
@@ -150,7 +154,7 @@ class AttendanceController extends Controller
             $return['status'] = 'warning';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Employee holiday has exists.';
-        }  else{
+        } else {
             $return['status'] = 'error';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Something goes to wrong';
@@ -159,7 +163,8 @@ class AttendanceController extends Controller
         exit;
     }
 
-    public function empSaveAdd(Request $request){
+    public function empSaveAdd(Request $request)
+    {
         $date = $request['date'];
         $objAttendance = new Attendance();
         $result = $objAttendance->empSaveAdd($request);
@@ -168,7 +173,7 @@ class AttendanceController extends Controller
             $return['status'] = 'success';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Employee Attendance details successfully added.';
-            $return['redirect'] = route('admin.attendance.day-list', ['date' => $date] );
+            $return['redirect'] = route('admin.attendance.day-list', ['date' => $date]);
         } elseif ($result == "attendance_exists") {
             $return['status'] = 'warning';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
@@ -177,7 +182,7 @@ class AttendanceController extends Controller
             $return['status'] = 'warning';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Employee holiday has exists.';
-        }  else{
+        } else {
             $return['status'] = 'error';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Something goes to wrong';
@@ -226,7 +231,8 @@ class AttendanceController extends Controller
         );
         return view('backend.pages.attendance.attendance_day_edit', $data);
     }
-    public function daySaveEdit(Request $request){
+    public function daySaveEdit(Request $request)
+    {
         $data['date'] = $request->date;
         $objAttendance = new Attendance();
         $result = $objAttendance->daySaveEdit($request);
@@ -234,7 +240,7 @@ class AttendanceController extends Controller
             $return['status'] = 'success';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
             $return['message'] = 'Attendance details successfully updated.';
-            $return['redirect'] = route('admin.attendance.day-list',['date'=>$data['date']]);
+            $return['redirect'] = route('admin.attendance.day-list', ['date' => $data['date']]);
         } elseif ($result == "Attendance_exists") {
             $return['status'] = 'error';
             $return['jscode'] = '$(".submitbtn:visible").removeAttr("disabled");$("#loader").hide();';
@@ -247,20 +253,68 @@ class AttendanceController extends Controller
         echo json_encode($return);
         exit;
     }
-    public function ajaxcall(Request $request){
+
+    public function showDeletedData(Request $request)
+    {
+
+        $data['title'] = Config::get('constants.PROJECT_NAME') . ' || Attendance List';
+        $data['description'] = Config::get('constants.PROJECT_NAME') . ' || Attendance List';
+        $data['keywords'] = Config::get('constants.PROJECT_NAME') . ' || Attendance List';
+        $data['css'] = array(
+            'toastr/toastr.min.css'
+        );
+        $data['plugincss'] = array(
+            'plugins/custom/datatables/datatables.bundle.css',
+            'plugins/custom/fullcalendar/fullcalendar.bundle.css',
+        );
+        $data['pluginjs'] = array(
+            'toastr/toastr.min.js',
+            'plugins/custom/datatables/datatables.bundle.js',
+            'pages/crud/datatables/data-sources/html.js',
+            'validate/jquery.validate.min.js',
+            'plugins/custom/fullcalendar/fullcalendar.bundle.js',
+            'pages/crud/forms/widgets/select2.js',
+        );
+        $data['js'] = array(
+            'comman_function.js',
+            'ajaxfileupload.js',
+            'jquery.form.min.js',
+            'attendance.js',
+        );
+        $data['funinit'] = array(
+            'Attendance.trash_init()',
+        );
+        $data['header'] = array(
+            'title' => 'Attendance Deleted List',
+            'breadcrumb' => array(
+                'Dashboard' => route('my-dashboard'),
+                'Attendance Deleted List' => 'Attendance Deleted List',
+            )
+        );
+        return view('backend.pages.attendance.trash', $data);
+    }
+
+    public function ajaxcall(Request $request)
+    {
         $action = $request->input('action');
         switch ($action) {
             case 'get_employee_list':
                 $data = $request->input('data');
                 $objEmployee = new Employee();
                 $data['employeeList'] = $objEmployee->get_admin_employee_attendance_details();
-                $details =  view('backend.pages.attendance.addEmployee', $data);
+                $details = view('backend.pages.attendance.addEmployee', $data);
                 echo $details;
                 break;
 
             case 'getdatatable':
                 $objAttendance = new Attendance();
                 $list = $objAttendance->getdatatable($request->input('data'));
+                echo json_encode($list);
+                break;
+
+            case 'get-attendance-data':
+                $objAttendance = new Attendance();
+                $list = $objAttendance->getAttendanceDatatable($request->input('data'));
                 echo json_encode($list);
                 break;
 
@@ -274,31 +328,39 @@ class AttendanceController extends Controller
                 $data = $request->input('data');
                 $objAttendance = new Attendance();
                 $list = $objAttendance->get_admin_attendance_details_by_day($request->input('data'));
-                $details =  view('backend.pages.attendance.attendance_day_list');
+                $details = view('backend.pages.attendance.attendance_day_list');
                 echo $details;
                 break;
 
             case 'common-activity':
                 $data = $request->input('data');
                 $objAttendance = new Attendance();
-
+                
                 $result = $objAttendance->common_activity($data);
                 if ($result) {
                     $return['status'] = 'success';
-                    if($data['dataAttr'] == 'emp_attendance'){
+                    if (isset($data['dataAttr'])) {
+
+                        if ($data['dataAttr'] == 'emp_attendance') {
+                            $attendanceId = $data['id'];
+                            $attendanceRecord = Attendance::find($attendanceId);
+                            $date = date_formate($attendanceRecord['date']);
+                            $return['message'] = 'Attendance details successfully deleted.';
+
+                        } else if ($data['dataAttr'] == 'emp_overtime') {
+                            $empOvertimeId = $data['id'];
+                            $empOvertimeRecord = EmpOvertime::find($empOvertimeId);
+                            $date = date_formate($empOvertimeRecord['date']);
+                            $return['message'] = 'Employee Overtime details successfully deleted.';
+                            $empOvertimeRecord->delete();
+                        }
+                    }
+
+                    if ($data['activity'] == 'restore-records') {
                         $attendanceId = $data['id'];
                         $attendanceRecord = Attendance::find($attendanceId);
                         $date = date_formate($attendanceRecord['date']);
-                        $return['message'] = 'Attendance details successfully deleted.';
-                        $attendanceRecord->delete();
-
-                    }
-                    else if($data['dataAttr'] == 'emp_overtime'){
-                        $empOvertimeId = $data['id'];
-                        $empOvertimeRecord = EmpOvertime::find($empOvertimeId);
-                        $date = date_formate($empOvertimeRecord['date']);
-                        $return['message'] = 'Employee Overtime details successfully deleted.';
-                        $empOvertimeRecord->delete();
+                        $return['message'] = "Attendance details successfully restore.";
                     }
 
                     $return['redirect'] = route('admin.attendance.day-list', ["date" => $date]);
@@ -306,7 +368,7 @@ class AttendanceController extends Controller
                 } else {
                     $return['status'] = 'error';
                     $return['jscode'] = '$("#loader").hide();';
-                    $return['message'] = 'It seems like something is wrong';;
+                    $return['message'] = 'It seems like something is wrong';
                 }
 
                 echo json_encode($return);
